refactor(cli): extract TypeScript plugin to separate package
