+++ .changeset/loud-chefs-stare.md
@@ -1,0 +1,5 @@
+---
+"@rnx-kit/cli": patch
+---
+
+Extract TypeScript plugin to a separate package
