+++ .changeset/odd-snails-unite.md
@@ -1,0 +1,5 @@
+---
+"@rnx-kit/metro-plugin-typescript": minor
+---
+
+Extract TypeScript plugin from `@rnx-kit/cli`
