+++ packages/cli/package.json
@@ -24,15 +24,14 @@
     "@rnx-kit/console": "^1.0.11",
     "@rnx-kit/metro-plugin-cyclic-dependencies-detector": "^1.0.21",
     "@rnx-kit/metro-plugin-duplicates-checker": "^2.0.0",
+    "@rnx-kit/metro-plugin-typescript": "^0.0.1",
     "@rnx-kit/metro-serializer": "^1.0.11",
     "@rnx-kit/metro-serializer-esbuild": "^0.1.0",
     "@rnx-kit/metro-service": "^3.0.2",
     "@rnx-kit/third-party-notices": "^1.2.13",
     "@rnx-kit/tools-language": "^1.4.1",
-    "@rnx-kit/tools-node": "^1.2.7",
-    "@rnx-kit/tools-react-native": "^1.2.0",
-    "@rnx-kit/typescript-react-native-resolver": "^0.2.0",
-    "@rnx-kit/typescript-service": "^1.5.3",
+    "@rnx-kit/tools-node": "^1.3.1",
+    "@rnx-kit/tools-react-native": "^1.2.3",
     "chalk": "^4.1.0",
     "find-up": "^5.0.0",
     "fs-extra": "^10.0.0",
@@ -62,8 +61,7 @@
     "jest": "^27.0.0",
     "jest-extended": "^2.0.0",
     "memfs": "^3.4.1",
-    "metro-config": "^0.67.0",
-    "typescript": "^4.0.0"
+    "metro-config": "^0.67.0"
   },
   "depcheck": {
     "ignoreMatches": [
