+++ packages/metro-plugin-typescript/src/index.ts
@@ -1,0 +1,2 @@
+export { createProjectCache } from "./projectCache";
+export { TypeScriptPlugin } from "./serializerHook";
