+++ packages/metro-plugin-typescript/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@rnx-kit/scripts/tsconfig-shared.json",
+  "include": ["src"]
+}
